+++ src/Resources/app/administration/src/module/paynl-settings/components/paynl-plugin-settings/index.js
@@ -1,4 +1,5 @@
 const { Component, Mixin } = Shopware;
+const { object, types } = Shopware.Utils;
 
 import template from './paynl-plugin-settings.html.twig';
 
@@ -6,10 +7,16 @@ Component.register('paynl-plugin-settings', {
     template,
 
     mixins: [
-        Mixin.getByName('notification')
+        Mixin.getByName('notification'),
+        Mixin.getByName('sw-inline-snippet')
     ],
 
-    inject: [ 'PaynlPaymentService' ],
+    inject: [
+        'PaynlPaymentService',
+        'acl'
+    ],
+
+    props: ['disabled'],
 
     data() {
         return {
@@ -41,7 +48,8 @@ Component.register('paynl-plugin-settings', {
                 usePAYStyles: null,
                 showDescription: null,
                 additionalAddressFields: null,
-                femaleSalutations: null
+                femaleSalutations: null,
+                paymentScreenLanguage: null,
             }
         };
     },
@@ -55,7 +63,19 @@ Component.register('paynl-plugin-settings', {
     computed: {
         credentialsEmpty: function() {
             return !this.tokenCodeFilled || !this.apiTokenFilled || !this.serviceIdFilled;
+        },
+
+        isDisabled: function () {
+            return this.isLoading || !this.acl.can('paynl.editor');
         }
+    },
+
+    mounted() {
+        this.$nextTick(function () {
+            if (!this.acl.can('paynl.editor')) {
+                this.disableSalesChannel();
+            }
+        })
     },
 
     methods: {
@@ -65,6 +85,29 @@ Component.register('paynl-plugin-settings', {
 
         installFinish() {
             this.isInstallSuccessful = false;
+        },
+
+        disableSalesChannel() {
+            var $this = this;
+            var waitSalesChannelInit = setInterval(function() {
+                let systemConfig = $this.$refs.systemConfig;
+                if (systemConfig === undefined
+                    || systemConfig.$children === undefined
+                    || systemConfig.$children[0] === undefined
+                ) {
+                    return;
+                }
+
+                let systemConfigSettings = systemConfig.$children[0];
+                if (systemConfigSettings.$children === undefined || systemConfigSettings.$children[0] === undefined) {
+                    return;
+                }
+
+                let salesChannelSwitch = systemConfigSettings.$children[0];
+                salesChannelSwitch.disabled = true;
+
+                clearInterval(waitSalesChannelInit);
+            }, 100); // check every 100ms
         },
 
         onConfigChange(config) {
@@ -84,6 +127,7 @@ Component.register('paynl-plugin-settings', {
                 additionalAddressFields: this.config['PaynlPaymentShopware6.settings.additionalAddressFields'],
                 femaleSalutations: this.config['PaynlPaymentShopware6.settings.femaleSalutations'],
                 usePAYStyles: this.config['PaynlPaymentShopware6.settings.usePAYStyles'],
+                paymentScreenLanguage: this.config['PaynlPaymentShopware6.settings.paymentScreenLanguage'],
             };
 
             this.showCredentilasErrors = false;
@@ -234,6 +278,57 @@ Component.register('paynl-plugin-settings', {
             }
 
             return element;
-        }
+        },
+
+        bindOriginalField(element, config) {
+            let originalElement;
+
+            element = this.bindField(element, config);
+
+            this.$refs.systemConfig.config.forEach((configElement) => {
+                configElement.elements.forEach((child) => {
+                    if (child.name === element.name) {
+                        originalElement = child;
+                        return;
+                    }
+                });
+            });
+
+            return originalElement || element;
+        },
+
+        getElementBind(element) {
+            const bind = object.deepCopyObject(element);
+
+            // Add inherited values
+            if (this.currentSalesChannelId !== null
+                && this.inherit
+                && this.actualConfigData.hasOwnProperty('null')
+                && this.actualConfigData.null[bind.name] !== null) {
+                if (bind.type === 'single-select' || bind.config.componentName === 'sw-entity-single-select') {
+                    // Add inherited placeholder option
+                    bind.placeholder = this.$tc('sw-settings.system-config.inherited');
+                } else if (bind.type === 'bool') {
+                    // Add inheritedValue for checkbox fields to restore the inherited state
+                    bind.config.inheritedValue = this.actualConfigData.null[bind.name] || false;
+                } else if (bind.type === 'password') {
+                    // Add inherited placeholder and mark placeholder as password so the rendering element
+                    // can choose to hide it
+                    bind.placeholderIsPassword = true;
+                    bind.placeholder = `${this.actualConfigData.null[bind.name]}`;
+                } else if (bind.type !== 'multi-select' && !types.isUndefined(this.actualConfigData.null[bind.name])) {
+                    // Add inherited placeholder
+                    bind.placeholder = `${this.actualConfigData.null[bind.name]}`;
+                }
+            }
+
+            // Add select properties
+            if (['single-select', 'multi-select'].includes(bind.type)) {
+                bind.config.labelProperty = 'name';
+                bind.config.valueProperty = 'id';
+            }
+
+            return bind;
+        },
     }
 });