+++ src/Resources/app/storefront/src/paynl-payment-plugin/paynl-payment-plugin.plugin.js
@@ -1,5 +1,4 @@
 import Plugin from 'src/plugin-system/plugin.class';
-import Datepicker from '../../node_modules/vanillajs-datepicker/js/Datepicker';
 
 export default class PaynlPaymentPlugin extends Plugin {
     init() {
@@ -10,15 +9,6 @@ export default class PaynlPaymentPlugin extends Plugin {
         const trigger = document.getElementById('paynl-payment-plugin');
 
         if (trigger) {
-            const elements = document.querySelectorAll('.paynl-dob');
-            Object.keys(elements).forEach(function(key) {
-                return new Datepicker(elements[key], {
-                    format: 'dd-mm-yyyy',
-                    autohide: true,
-                    maxDate: new Date(),
-                });
-            });
-
             const form = trigger.parentNode;
             form.addEventListener('submit', this.onSavePaymentMethod);
             form.addEventListener('change', this.onChangeCallback);
@@ -78,7 +68,7 @@ export default class PaynlPaymentPlugin extends Plugin {
 
     onChangeCallback(event) {
         // check if event is triggered by payment method change
-        if (event.target.classList.contains('payment-method-input')) {
+        if (event.target.className.indexOf('payment-method-input') !== -1) {
             // hide previous extra data block
             const prevExtraDataBlock = event.currentTarget.querySelector('.paynl-payment-method-extra.active');
             if (prevExtraDataBlock) {
@@ -87,11 +77,8 @@ export default class PaynlPaymentPlugin extends Plugin {
 
             const currentPaymentBlock = event.target.parentNode.parentNode.parentNode;
             const extraDataBlock = currentPaymentBlock.querySelector('.paynl-payment-method-extra');
-
-            // set issuer value to initial position
-            const issuerSelect = extraDataBlock.querySelector('.paynl-ideal-banks-select');
-            if (issuerSelect) {
-                issuerSelect.value = '';
+            if (!extraDataBlock) {
+                return;
             }
 
             const paymentMethodBankName = event.currentTarget.querySelector('#paymentMethodBankName');