+++ src/Storefront/Controller/PaynlAccountOrderController.php
@@ -48,7 +48,6 @@ class PaynlAccountOrderController extends StorefrontController
         $salesChannelContext = $request->attributes->get(PlatformRequest::ATTRIBUTE_SALES_CHANNEL_CONTEXT_OBJECT);
         $dob = $request->get('dob');
         $phone = $request->get('phone');
-        $issuer = $request->get('issuer');
         /** @var CustomerEntity $customer */
         $customer = $salesChannelContext->getCustomer();
         /** @var CustomerAddressEntity $billingAddress */
@@ -60,12 +59,6 @@ class PaynlAccountOrderController extends StorefrontController
 
         if (!empty($phone)) {
             $this->customerHelper->saveCustomerPhone($billingAddress, $phone, $context);
-        }
-
-        if (!empty($issuer)) {
-            $this->session->set('paynlIssuer', $issuer);
-        } else {
-            $this->session->remove('paynlIssuer');
         }
 
         return new JsonResponse(['success' => true]);